+++ tools/src/com.oracle.truffle.tools.profiler/src/com/oracle/truffle/tools/profiler/impl/CPUSamplerCLI.java
@@ -42,7 +42,6 @@ import org.graalvm.options.OptionType;
 import org.graalvm.options.OptionValues;
 
 import com.oracle.truffle.api.Option;
-import com.oracle.truffle.api.TruffleContext;
 import com.oracle.truffle.api.instrumentation.TruffleInstrument;
 import com.oracle.truffle.tools.profiler.CPUSampler;
 import com.oracle.truffle.tools.profiler.CPUSamplerData;
@@ -167,7 +166,7 @@ class CPUSamplerCLI extends ProfilerCLI {
 
     static void handleOutput(TruffleInstrument.Env env, CPUSampler sampler) {
         PrintStream out = chooseOutputStream(env, OUTPUT_FILE);
-        Map<TruffleContext, CPUSamplerData> data = sampler.getData();
+        List<CPUSamplerData> data = sampler.getDataList();
         OptionValues options = env.getOptions();
         switch (options.get(OUTPUT)) {
             case HISTOGRAM:
@@ -186,15 +185,15 @@ class CPUSamplerCLI extends ProfilerCLI {
         }
     }
 
-    private static void printSamplingCallTree(PrintStream out, OptionValues options, Map<TruffleContext, CPUSamplerData> data) {
-        for (Map.Entry<TruffleContext, CPUSamplerData> entry : data.entrySet()) {
-            new SamplingCallTree(entry.getValue(), options).print(out);
+    private static void printSamplingCallTree(PrintStream out, OptionValues options, List<CPUSamplerData> data) {
+        for (CPUSamplerData entry : data) {
+            new SamplingCallTree(entry, options).print(out);
         }
     }
 
-    private static void printSamplingHistogram(PrintStream out, OptionValues options, Map<TruffleContext, CPUSamplerData> data) {
-        for (Map.Entry<TruffleContext, CPUSamplerData> entry : data.entrySet()) {
-            new SamplingHistogram(entry.getValue(), options).print(out);
+    private static void printSamplingHistogram(PrintStream out, OptionValues options, List<CPUSamplerData> data) {
+        for (CPUSamplerData entry : data) {
+            new SamplingHistogram(entry, options).print(out);
         }
     }
 
@@ -217,7 +216,7 @@ class CPUSamplerCLI extends ProfilerCLI {
     }
 
     private static boolean sampleDurationTooLong(CPUSampler sampler) {
-        for (CPUSamplerData value : sampler.getData().values()) {
+        for (CPUSamplerData value : sampler.getDataList()) {
             if (value.getSampleDuration().getAverage() > MAX_OVERHEAD_WARNING_THRESHOLD * sampler.getPeriod() * MILLIS_TO_NANOS) {
                 return true;
             }
@@ -229,13 +228,13 @@ class CPUSamplerCLI extends ProfilerCLI {
         out.println("-------------------------------------------------------------------------------- ");
     }
 
-    private static void printSamplingJson(PrintStream out, OptionValues options, Map<TruffleContext, CPUSamplerData> data) {
+    private static void printSamplingJson(PrintStream out, OptionValues options, List<CPUSamplerData> data) {
         boolean gatheredHitTimes = options.get(GATHER_HIT_TIMES);
         JSONObject output = new JSONObject();
         output.put("tool", CPUSamplerInstrument.ID);
         output.put("version", CPUSamplerInstrument.VERSION);
         JSONArray contexts = new JSONArray();
-        for (CPUSamplerData samplerData : data.values()) {
+        for (CPUSamplerData samplerData : data) {
             contexts.put(perContextData(samplerData, gatheredHitTimes));
         }
         output.put("contexts", contexts);